rename subdir py_gtfs_rt_ingestion to lib
