+++ py_gtfs_rt_ingestion/lib/config_rt_vehicle.py
