+++ py_gtfs_rt_ingestion/lib/gtfs_rt_detail.py
